chore: tag 엔티티 별도 모듈로 분리 boostcamp#54
