+++ backend/src/diaries/diaries.module.ts
@@ -2,10 +2,11 @@ import { Module } from '@nestjs/common';
 import { DiariesController } from './diaries.controller';
 import { DiariesService } from './diaries.service';
 import { DiariesRepository } from './diaries.repository';
-import { TagsRepository } from './tag.repository';
+import { TagsModule } from 'src/tags/tags.module';
 
 @Module({
   controllers: [DiariesController],
-  providers: [DiariesService, DiariesRepository, TagsRepository],
+  providers: [DiariesService, DiariesRepository],
+  imports: [TagsModule],
 })
 export class DiariesModule {}
