+++ backend/src/diaries/diaries.service.ts
@@ -1,34 +1,19 @@
 import { Injectable } from '@nestjs/common';
 import { DiariesRepository } from './diaries.repository';
 import { CreateDiaryDto } from './dto/diary.dto';
-import { TagsRepository } from './tag.repository';
 import { User } from 'src/users/entity/user.entity';
+import { TagsService } from 'src/tags/tags.service';
 
 @Injectable()
 export class DiariesService {
   constructor(
     private readonly diariesRepository: DiariesRepository,
-    private readonly tagsRepository: TagsRepository,
+    private readonly tagsService: TagsService,
   ) {}
 
   async saveDiary(user: User, createDiaryDto: CreateDiaryDto) {
-    const tags = await this.mapTagNameToTagType(createDiaryDto.tagNames);
+    const tags = await this.tagsService.mapTagNameToTagType(createDiaryDto.tagNames);
 
     await this.diariesRepository.saveDiary(user, createDiaryDto, tags);
-  }
-
-  async mapTagNameToTagType(tagNames: string[]) {
-    if (!tagNames) return null;
-
-    return await Promise.all(
-      tagNames.map(async (tagName) => {
-        let tag = await this.tagsRepository.findByName(tagName);
-
-        if (!tag) {
-          tag = await this.tagsRepository.saveTag(tagName);
-        }
-        return tag;
-      }),
-    );
   }
 }
