+++ backend/src/diaries/entity/diary.entity.ts
@@ -12,7 +12,7 @@ import {
   UpdateDateColumn,
 } from 'typeorm';
 import { DiaryStatus } from './diaryStatus';
-import { Tag } from './tag.entity';
+import { Tag } from '../../tags/entity/tag.entity';
 
 @Entity()
 export class Diary extends BaseEntity {
